refactor: split koan-music into koan-tui + koan-server + koan-cli (#157)
Compiler-enforced crate boundaries for portability. koan-core's pub
surface is the contract — downstream crates can only use what's
explicitly public. Future iOS app imports only koan-core.

Four crates:
- koan-core: audio engine, player, DB, helpers (subsonic client,
  cache paths, track resolution, download). Platform-agnostic.
- koan-tui: Ratatui TUI, visualizers, media keys, download queue.
  Exports run_tui(). Library crate.
- koan-server: GraphQL + Subsonic REST + MCP server. Library crate.
- koan-cli: clap CLI, logger, signal handling. Produces `koan` binary.

Dependency rules enforced by Cargo:
  koan-cli  → koan-core + koan-tui + koan-server
  koan-tui  → koan-core (NOT koan-server)
  koan-server → koan-core (NOT koan-tui)

All 598 tests pass. Zero clippy warnings.

Closes #157
