+++ crates/koan-cli/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "koan-cli"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "CLI entry point for koan — bit-perfect music player."
+keywords = ["audio", "music", "player", "cli"]
+categories = ["multimedia::audio"]
+
+[[bin]]
+name = "koan"
+path = "src/main.rs"
+
+[dependencies]
+chrono = "0.4"
+clap = { version = "4.5", features = ["derive"] }
+clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
+crossbeam-channel = "0.5"
+crossterm = "0.28"
+ctrlc = "3"
+jwalk = "0.8"
+koan-core = { path = "../koan-core", version = "0.20.4" }
+koan-server = { path = "../koan-server", version = "0.20.4" }
+koan-tui = { path = "../koan-tui", version = "0.20.4" }
+log = "0.4"
+owo-colors = "4"
+rayon = "1"
+rpassword = "7.4"
+toml = "0.9"
+walkdir = { workspace = true }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9"
+
+[dev-dependencies]
+tempfile = "3"
