+++ crates/koan-cli/src/commands/analyze.rs
