+++ crates/koan-cli/src/commands/cache.rs
