+++ crates/koan-cli/src/commands/probe.rs
