+++ crates/koan-cli/src/commands/scan.rs
