+++ crates/koan-cli/src/commands/search.rs
