+++ crates/koan-core/src/lib.rs
@@ -4,6 +4,7 @@ pub mod credentials;
 pub mod db;
 pub mod format;
 pub mod graphql_client;
+pub mod helpers;
 pub mod index;
 pub mod lyrics;
 pub mod organize;
