--- crates/koan-music/Cargo.toml
@@ -1,53 +1,0 @@
-[package]
-name = "koan-music"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Bit-perfect music player — Ratatui TUI, gapless playback, Navidrome/Subsonic integration."
-keywords = ["audio", "music", "player", "tui"]
-categories = ["multimedia::audio"]
-
-[[bin]]
-name = "koan"
-path = "src/main.rs"
-
-[dependencies]
-chrono = "0.4"
-clap = { version = "4.5", features = ["derive"] }
-clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
-ctrlc = "3"
-crossbeam-channel = "0.5"
-crossterm = "0.28"
-image = { version = "0.25", default-features = false, features = ["jpeg", "png", "webp"] }
-koan-core = { path = "../koan-core", version = "0.20.4" }
-log = "0.4"
-nucleo = "0.5"
-owo-colors = "4"
-ratatui = "0.29"
-jwalk = "0.8"
-rayon = "1"
-rpassword = "7.4"
-souvlaki = "0.8"
-toml = "0.9"
-walkdir = { workspace = true }
-rmcp = { version = "1.2.0", features = ["server", "macros", "transport-async-rw", "transport-io"] }
-tokio = { version = "1.50.0", features = ["rt-multi-thread", "net", "macros", "signal"] }
-serde_json = "1.0.149"
-serde = { version = "1.0.228", features = ["derive"] }
-uuid = { version = "1.22.0", features = ["v7"] }
-async-graphql = "7"
-async-graphql-axum = "7"
-axum = "0.8"
-base64 = "0.22.1"
-md5 = "0.8"
-rusqlite = { workspace = true }
-tokio-util = { version = "0.7", features = ["io"] }
-reqwest = { version = "0.13.2", features = ["stream"] }
-tower = { version = "0.5", features = ["util", "limit", "load-shed"] }
-
-[target.'cfg(target_os = "macos")'.dependencies]
-core-foundation = "0.9"
-
-[dev-dependencies]
-tempfile = "3"
