--- crates/koan-music/src/tui/mod.rs
@@ -1,17 +1,0 @@
-pub mod app;
-pub mod context_menu;
-pub mod cover_art;
-pub mod device_selector;
-pub mod help_modal;
-pub mod keys;
-pub mod library;
-pub mod lyrics;
-pub mod organize;
-pub mod picker;
-pub mod queue;
-pub mod theme;
-pub mod track_info;
-pub mod transport;
-pub mod ui;
-pub mod visualizer;
-pub mod viz_picker;
