+++ crates/koan-tui/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "koan-tui"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Ratatui TUI for koan music player. Visualizers, media keys, transport."
+keywords = ["audio", "music", "player", "tui"]
+categories = ["multimedia::audio"]
+
+[dependencies]
+chrono = "0.4"
+crossbeam-channel = "0.5"
+crossterm = "0.28"
+image = { version = "0.25", default-features = false, features = ["jpeg", "png", "webp"] }
+jwalk = "0.8"
+koan-core = { path = "../koan-core", version = "0.20.4" }
+koan-server = { path = "../koan-server", version = "0.20.4" }
+log = "0.4"
+reqwest = { version = "0.13.2", default-features = false, features = ["blocking", "json", "rustls"] }
+nucleo = "0.5"
+owo-colors = "4"
+ratatui = "0.29"
+rayon = "1"
+souvlaki = "0.8"
+toml = "0.9"
+uuid = { version = "1.22.0", features = ["v7"] }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9"
+
+[dev-dependencies]
+tempfile = "3"
