+++ crates/koan-tui/src/context_menu.rs
