+++ crates/koan-tui/src/cover_art.rs
