+++ crates/koan-tui/src/device_selector.rs
