+++ crates/koan-tui/src/help_modal.rs
