+++ crates/koan-tui/src/lyrics.rs
