+++ crates/koan-tui/src/organize.rs
