+++ crates/koan-tui/src/picker.rs
