+++ crates/koan-tui/src/transport.rs
