+++ crates/koan-tui/src/viz_picker.rs
